Second disk drive image (B:) supported
Issue #2

The optional disk image file for drive B: is named cpmdisk2.bin. Be
sure to update the file system.bin. See the appended UPDATE section
in INSTALL or INSTALL.linux for details!

The contents of the RAM disk(s) is saved on shutdown only when the
drive has been written now.